@@ -6,79 +6,85 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "start": "node tools/start.js",
-    "build": "node tools/build.js",
-    "eslint": "eslint --ext .jsx,.js src/",
+    "start": "babel-node tools/start.js",
+    "build": "babel-node tools/build.js",
+    "eslint": "esw src tools --color",
     "stylelint": "stylelint 'src/**/*.css'"
   },
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-react-css-modules": "^3.1.0",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "babel-runtime": "^6.23.0",
-    "browser-sync": "^2.18.8",
-    "chalk": "^2.0.1",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.4",
-    "doiuse": "^3.0.0",
-    "eslint": "^4.2.0",
-    "eslint-config-airbnb-base": "^11.2.0",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "favicons-webpack-plugin": "0.0.7",
-    "file-loader": "^0.11.2",
-    "html-webpack-plugin": "^2.29.0",
-    "i18next": "^8.4.3",
-    "i18next-xhr-backend": "^1.4.2",
-    "normalize.css": "^6.0.0",
-    "postcss-assets": "^4.2.0",
-    "postcss-center": "^1.0.0",
-    "postcss-clearfix": "^2.0.0",
-    "postcss-cssnext": "^2.11.0",
-    "postcss-import": "^9.1.0",
-    "postcss-inline-svg": "^2.3.1",
-    "postcss-loader": "^2.0.6",
-    "postcss-modules": "^0.8.0",
-    "postcss-reporter": "^4.0.0",
-    "postcss-short": "^3.0.3",
-    "postcss-use": "^2.3.0",
-    "react-hot-loader": "^3.0.0-beta.7",
-    "redux-devtools-extension": "^2.13.2",
-    "redux-logger": "^3.0.6",
-    "style-loader": "^0.18.2",
-    "stylelint": "^8.0.0",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.1",
+    "babel-jest": "^22.1.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-react-constant-elements": "^6.23.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.12",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-stage-1": "^6.24.1",
+    "browser-sync": "^2.23.5",
+    "connect-history-api-fallback": "^1.5.0",
+    "eslint": "^4.15.0",
+    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-react": "^7.5.1",
+    "eslint-watch": "^3.1.3",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "html-webpack-plugin": "^2.30.1",
+    "react-hot-loader": "^3.1.3",
+    "stylelint": "^8.4.0",
     "stylelint-config-css-modules": "^1.1.0",
-    "stylelint-config-standard": "^17.0.0",
-    "stylelint-formatter-pretty": "^1.0.3",
-    "stylelint-webpack-plugin": "^0.9.0",
-    "uglifyjs-webpack-plugin": "^0.4.6",
-    "url-loader": "^0.5.9",
-    "webpack": "^3.3.0",
-    "webpack-dev-middleware": "^1.11.0",
-    "webpack-hot-middleware": "^2.18.2",
-    "webpack-md5-hash": "0.0.5"
+    "stylelint-config-standard": "^18.0.0",
+    "webpack": "^3.10.0",
+    "webpack-dev-middleware": "^2.0.4",
+    "webpack-hot-middleware": "^2.21.0",
+    "webpack-md5-hash": "^0.0.6"
+  },
+  "babel": {
+    "env": {
+      "development": {
+        "presets": [
+          "env",
+          "react",
+          "react-hmre",
+          "stage-1"
+        ]
+      },
+      "production": {
+        "presets": [
+          [
+            "env",
+            {
+              "es2015": {
+                "modules": false
+              },
+              "targets": {
+                "ie": 9,
+                "uglify": true
+              }
+            }
+          ],
+          "react",
+          "stage-1"
+        ],
+        "plugins": [
+          "transform-react-constant-elements",
+          "transform-react-remove-prop-types"
+        ]
+      },
+      "test": {
+        "presets": [
+          "env",
+          "react",
+          "stage-1"
+        ]
+      }
+    }
   },
   "dependencies": {
-    "i18next-browser-languagedetector": "^2.0.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-i18next": "^4.6.4",
-    "react-redux": "^5.0.5",
-    "redux": "^3.7.2",
-    "redux-saga": "^0.15.4",
-    "redux-seamless-immutable": "^0.3.0",
-    "reset.css": "^2.0.2",
-    "seamless-immutable": "^7.1.2",
-    "whatwg-fetch": "^2.0.3"
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0"
   }
 }
