@@ -1,28 +1,24 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import { I18nextProvider } from 'react-i18next';
+// import { Provider } from 'react-redux';
+// import { I18nextProvider } from 'react-i18next';
 import { AppContainer } from 'react-hot-loader';
 
-import 'normalize.css';
-import 'reset.css';
-import './assets/main.css';
-import './assets/fonts/Roboto/stylesheet.css';
+// import 'normalize.css';
+// import 'reset.css';
+// import './assets/main.css';
+// import './assets/fonts/Roboto/stylesheet.css';
 
-import configureStore from './redux/configureStore';
-import i18next from './utils/i18next';
+// import configureStore from './redux/configureStore';
+// import i18next from './utils/i18next';
 import App from './containers/App';
 
-const store = configureStore({});
+// const store = configureStore({});
 
 const render = (Component) => {
   ReactDOM.render(
     <AppContainer>
-      <Provider store={store}>
-        <I18nextProvider i18n={i18next}>
-          <Component/>
-        </I18nextProvider>
-      </Provider>
+      <Component/>
     </AppContainer>,
     document.getElementById('app')
   );
